chore: update player timestamps types

+++ CityVilleDotnet.Domain/GameEntities/PlayerDto.cs
@@ -11,7 +11,7 @@ public class PlayerDto
     [JsonPropertyName("uid")] public string Uid { get; set; } = "333";
 
     [JsonPropertyName("lastTrackingTimestamp")]
-    public int LastTrackingTimestamp { get; set; } = 0;
+    public long LastTrackingTimestamp { get; set; } = 0;
 
     [JsonPropertyName("playerNews")] public List<object> PlayerNews { get; set; } = [];
